Update Process_Expression.java
Used stack functions: push(string), used(), String top(), pop().  Added more to Process_Expression.  Started to fill out inset to rpn conversion.  Also it might be best to solve the equation as it is converted.  Right after an operator is added, its expression can be evaluated and re-pushed.  Feel free to modify if something is wrong.